⬆️ add php 8.5

+++ Classes/Service/IpCountryLocator/MmdbFile.php
@@ -127,7 +127,7 @@ final class MmdbFile implements IpCountryLocatorInterface, SingletonInterface
     private function humanFilesize(int $bytes, int $decimals = 2): string
     {
         $size = ['B', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-        $factor = floor((strlen((string)$bytes) - 1) / 3);
+        $factor = (int)floor((strlen((string)$bytes) - 1) / 3);
         return sprintf(sprintf('%%.%df', $decimals), $bytes / (1024 ** $factor)) . ($size[$factor] ?? '');
     }
 }
